Barebones EFI implementation for mach-virt/arm64
This is a minimal implementation in Rust of those parts of the EFI spec
that are actually needed to boot a arm64 Linux kernel straight from QEMU's
mach-virt.

It uses the efiloader crate which implements the bare minimum of EFI's
boot and runtime services that are needed to boot a Linux kernel, UKI
image or GRUB/shim loader stage.

It is intended to be minimal and secure:
- file data is never copied around - the kernel image and initrd are
  DMA'd into their final location using QEMU's fwcfg device,
- the MMU is enabled straight out of reset, and all memory accesses and
  code execution occurs with the appropriate memory and permission
  attributes,
- written in Rust,
- only a reduced subset of boot and runtime services is implemented,
- ACPI and SMBIOS tables are obtained from QEMU fwcfg directly and not
  modified,

How to use:

- build using 'cargo build'
  (ELF file ends up in target/aarch64-unknown-none/.../efilite)

- use 'objcopy -O binary <in> <out>' to convert the ELF file into a raw binary

- run using something like the below

$ qemu-system-aarch64 \
    -M virt -cpu max -smp 4 \
    -net none -nographic -m 900m -bios efilite.bin -kernel path/to/Image \
    -drive if=virtio,file=path/to/hda.xxx,format=xxx -initrd <initrd>