Barebones EFI implementation for mach-virt/arm64
This is a minimal implementation in Rust of those parts of the EFI spec
that are actually needed to boot a arm64 Linux kernel straight from QEMU's
mach-virt.

It includes a protocol database, a memory map and the various bits and
pieces that the EFI stub as well as the core kernel rely on at runtime.

It is intended to be minimal and secure:
- code is never copied around - the kernel image and initrd are DMA'd
  into their final location using QEMU's fwcfg device,
- the MMU is enabled straight out of reset, and all memory accesses and
  code execution occurs with the appropriate memory and permission
  attributes,
- written in Rust, (*)
- only a minimal subset of boot and runtime services is implemented,

This is proof-of-concept code. Notable issues are:
- DT only
- QEMU's fwcfg device is the only supported device

How to use:

- build using 'cargo build'
  (ELF file ends up in target/aarch64-unknown-linux-gnu/release/efilite)

- use 'objcopy -O binary <in> <out>' to convert the ELF file into a raw binary

- run using something like the below

$ qemu-system-aarch64 \
    -M virt,gic-version=host -cpu host -enable-kvm -smp 4 \
    -net none -nographic -m 900m -bios efilite.bin -kernel path/to/Image \
    -drive if=virtio,file=path/to/hda.xxx,format=xxx -initrd <initrd>

(*) this is my first Rust program so there are undoubtedly issues that
    need fixing