Use RNDR instruction if available
Use the RNDR instruction to back get_random_u64() if it is implemented.

This permits QEMU running in TCG mode with -cpu max to use the
randomization feature, as it does not implement the TRNG hypercall.

While at it, rename get_random_bytes() to get_entropy() to emphasize
that it is backed only by the TRNG service and does not fallback to RNDR
as get_random_u64() now does.