Remove remaining spinlock uses
Using spinlocks in single threaded code does not make a lot of sense:
the code will either execute correctly, or hang when a reference to the
encapsulated object is taken, even if the reference is shared.

It is better to use RefCells for interior mutability, and trigger a
panic when several mutable borrows are taken.