Use addr_of_mut! macro to access mutable static memory pools
Rust 1.77 warns about taking mutable references to static mutable
objects, on the basis that it might result in UB if concurrent accesses
are made.

This cannot occur in these cases, as the mutable globals are local to
the scope in question, and so there is no way for other code to take
such a reference.

Adopt the workaround proposed by the compiler, which essentially papers
over the issue, but that is also fine for the same reasons stated above.